Improvement/authorization logic (#1517)
* New component for Authorization and removing authorization logic from the operation-console

* Using the new Authorization component in the graphql console and fixing some minor bugs

* Removing reference to gql from authorization and updating css classes

* css classes updates
